+++ src/lib/geomutil/crayplutil/crayList.c
@@ -104,44 +104,44 @@ static Geom *ListElement(Geom *list, int elem) {
  */
 #define LIST_EACH(list, func) \
 { \
-  int val; \
+  long val; \
   for (val = 0; (list = (Geom*)((List *)(list))->cdr);) \
-    val |= (int)func; \
+    val |= (long)func; \
   return (void *)val; \
 }
 
 void *cray_list_HasColor(int sel, Geom *geom, va_list *args) {
   int *gpath = va_arg(*args, int *);
   if (gpath != NULL) 
-    return (void *)(crayHasColor(ListElement(geom, *gpath), gpath + 1));
+    return (void *)(long)(crayHasColor(ListElement(geom, *gpath), gpath + 1));
   LIST_EACH(geom, crayHasColor(((List *)geom)->car, NULL));
 }
 
 void *cray_list_HasVColor(int sel, Geom *geom, va_list *args) {
   int *gpath = va_arg(*args, int *);
   if (gpath != NULL) 
-    return (void *)(crayHasVColor(ListElement(geom, *gpath), gpath + 1));
+    return (void *)(long)(crayHasVColor(ListElement(geom, *gpath), gpath + 1));
   LIST_EACH(geom, crayHasVColor(((List *)geom)->car, NULL));
 }
 
 void *cray_list_HasFColor(int sel, Geom *geom, va_list *args) {
   int *gpath = va_arg(*args, int *);
   if (gpath != NULL) 
-    return (void *)(crayHasFColor(ListElement(geom, *gpath), gpath + 1));
+    return (void *)(long)(crayHasFColor(ListElement(geom, *gpath), gpath + 1));
   LIST_EACH(geom, crayHasFColor(((List *)geom)->car, NULL));
 }
 
 void *cray_list_CanUseVColor(int sel, Geom *geom, va_list *args) {
   int *gpath = va_arg(*args, int *);
   if (gpath != NULL) 
-    return (void *)(crayCanUseVColor(ListElement(geom, *gpath), gpath + 1));
+    return (void *)(long)(crayCanUseVColor(ListElement(geom, *gpath), gpath + 1));
   LIST_EACH(geom, crayCanUseVColor(((List *)geom)->car, NULL));
 }
 
 void *cray_list_CanUseFColor(int sel, Geom *geom, va_list *args) {
   int *gpath = va_arg(*args, int *);
   if (gpath != NULL) 
-    return (void *)(crayCanUseFColor(ListElement(geom, *gpath), gpath + 1));
+    return (void *)(long)(crayCanUseFColor(ListElement(geom, *gpath), gpath + 1));
   LIST_EACH(geom, crayCanUseFColor(((List *)geom)->car, NULL));
 }
 
@@ -149,7 +149,7 @@ void *cray_list_UseVColor(int sel, Geom *geom, va_list *args) {
   ColorA *c = va_arg(*args, ColorA *);
   int *gpath = va_arg(*args, int *);
   if (gpath != NULL) 
-    return (void *)(crayUseVColor(ListElement(geom, *gpath), c, gpath + 1));
+    return (void *)(long)(crayUseVColor(ListElement(geom, *gpath), c, gpath + 1));
   LIST_EACH(geom, crayUseVColor(((List *)geom)->car, c, NULL));
 }
 
@@ -157,14 +157,14 @@ void *cray_list_UseFColor(int sel, Geom *geom, va_list *args) {
   ColorA *c = va_arg(*args, ColorA *);
   int *gpath = va_arg(*args, int *);
   if (gpath != NULL) 
-    return (void *)(crayUseFColor(ListElement(geom, *gpath), c, gpath + 1));
+    return (void *)(long)(crayUseFColor(ListElement(geom, *gpath), c, gpath + 1));
   LIST_EACH(geom, crayUseFColor(((List *)geom)->car, c, NULL));
 }
 
 void *cray_list_EliminateColor(int sel, Geom *geom, va_list *args) {
   int *gpath = va_arg(*args, int *);
   if (gpath != NULL) 
-    return (void *)(crayEliminateColor(ListElement(geom, *gpath), gpath + 1));
+    return (void *)(long)(crayEliminateColor(ListElement(geom, *gpath), gpath + 1));
   LIST_EACH(geom, crayEliminateColor(((List *)geom)->car, NULL));
 }
 
@@ -173,7 +173,7 @@ void *cray_list_SetColorAll(int sel, Geom *geom, va_list *args) {
   int *gpath = va_arg(*args, int *);
 
   if (gpath != NULL) 
-    return (void *)craySetColorAll(ListElement(geom, *gpath), c, gpath + 1);
+    return (void *)(long)craySetColorAll(ListElement(geom, *gpath), c, gpath + 1);
   LIST_EACH(geom, craySetColorAll(((List *)geom)->car, c, NULL));
 }
 
@@ -183,7 +183,7 @@ void *cray_list_SetColorAt(int sel, Geom *geom, va_list *args) {
   *edge = va_arg(*args, int *), *gpath = va_arg(*args, int *);
   HPoint3 *pt = va_arg(*args, HPoint3 *);
   if (gpath != NULL) 
-    return (void *)(craySetColorAt(ListElement(geom, *gpath), c, vindex,
+    return (void *)(long)(craySetColorAt(ListElement(geom, *gpath), c, vindex,
 				   findex, edge, gpath + 1, pt));
   LIST_EACH(geom, craySetColorAt(((List *)geom)->car, c, vindex, findex,
 				 edge, NULL, pt));
@@ -194,7 +194,7 @@ void *cray_list_SetColorAtV(int sel, Geom *geom, va_list *args) {
   int index = va_arg(*args, int), *gpath = va_arg(*args, int *);
   HPoint3 *pt = va_arg(*args, HPoint3 *);
   if (gpath != NULL) 
-    return (void *)(craySetColorAtV(ListElement(geom, *gpath), c, index, 
+    return (void *)(long)(craySetColorAtV(ListElement(geom, *gpath), c, index, 
 				    gpath + 1, pt));
   LIST_EACH(geom, craySetColorAtV(((List *)geom)->car, c, index, NULL, pt));
 }
@@ -203,7 +203,7 @@ void *cray_list_SetColorAtF(int sel, Geom *geom, va_list *args) {
   ColorA *c = va_arg(*args, ColorA *);
   int index = va_arg(*args, int), *gpath = va_arg(*args, int *);
   if (gpath != NULL) 
-    return (void *)(craySetColorAtF(ListElement(geom, *gpath), c, index, 
+    return (void *)(long)(craySetColorAtF(ListElement(geom, *gpath), c, index, 
 				    gpath + 1));
   LIST_EACH(geom, craySetColorAtF(((List *)geom)->car, c, index, NULL));
 }
@@ -214,7 +214,7 @@ void *cray_list_GetColorAt(int sel, Geom *geom, va_list *args) {
   *edge = va_arg(*args, int *), *gpath = va_arg(*args, int *);
   HPoint3 *pt = va_arg(*args, HPoint3 *);
   if (gpath != NULL) 
-    return (void *)(crayGetColorAt(ListElement(geom, *gpath), c, vindex, 
+    return (void *)(long)(crayGetColorAt(ListElement(geom, *gpath), c, vindex, 
 				   findex, edge, gpath + 1, pt));
   LIST_EACH(geom, crayGetColorAt(((List *)geom)->car, c, vindex, 
 				 findex, edge, NULL, pt));
@@ -225,7 +225,7 @@ void *cray_list_GetColorAtV(int sel, Geom *geom, va_list *args) {
   int index = va_arg(*args, int), *gpath = va_arg(*args, int *);
   HPoint3 *pt = va_arg(*args, HPoint3 *);
   if (gpath != NULL)
-      return (void *)(crayGetColorAtV(ListElement(geom, *gpath), c, index, 
+      return (void *)(long)(crayGetColorAtV(ListElement(geom, *gpath), c, index, 
 				      gpath + 1, pt));
   LIST_EACH(geom, crayGetColorAtV(((List *)geom)->car, c, index, NULL, pt));
 }
@@ -234,7 +234,7 @@ void *cray_list_GetColorAtF(int sel, Geom *geom, va_list *args) {
   ColorA *c = va_arg(*args, ColorA *);
   int index = va_arg(*args, int), *gpath = va_arg(*args, int *);
   if (gpath != NULL) 
-    return (void *)(crayGetColorAtF(ListElement(geom, *gpath), c, index, 
+    return (void *)(long)(crayGetColorAtF(ListElement(geom, *gpath), c, index, 
 				    gpath + 1));
   LIST_EACH(geom, crayGetColorAtF(((List *)geom)->car, c, index, NULL));
 }
